Fix injecting as Administrator on Windows 7 and newer
Making use of a Windows service whenever possible is not a good idea given
that Windows 7 and newer have a tighter policy for CreateRemoteThread.
We now only try the service approach as a last resort, which is when the
target process actually is a Windows service.

Thanks to markwo and egirault for helping track down this issue.

Fixes #2