Fix exception type thrown by interfaces exposed by proxies
We cannot declare them as throwing frida.Error due to the fact that
Vala's GDBus code generator produces proxy objects that also throw
transport errors (IOError). Because of this constraint we have to
declare these methods as throwing GLib.Error to cover the transport
error cases. Otherwise we'll end up with failing runtime assertions
when suddenly an IOError appears instead of a frida.Error.