Fix agent regression on Darwin
We cannot use `RTLD_NEXT` when we are calling from an image that we
manually mapped into memory, and consequently dyld knows nothing about.
As we don't actually need to resolve the address anyway we can do it the
way we used to on Darwin and leave the `dlsym()` for other UNIXes.