Prevent libdl from unloading to work around TLS destructor bug
Keep a handle to libdl.so.2 so that it is never unloaded even when
libfrida-agent.so unloads in a process that did not previously have
libdl loaded.

This is to prevent a bug which was encountered where libdl tries to
perform a TLS key destructor after the libdl code has been unloaded,
resulting in a segfualt at thread cleanup time.