Add support for multiple agent sessions (#85)
This addresses a long-standing design flaw where each `attach()` request would
return the same object, or more precisely, the same shared session. The result
was that clients would need to coordinate so that two clients interested in
attaching to the same process wouldn't step on each other in case one decided
to detach() once it was done.