Fix stale injection state after an exec transition
We read back the `stay_resident` value when the remote POSIX thread
terminates. This is however very risky in case the process is being
replaced, for example through posix_spawn(POSIX_SPAWN_SETEXEC), as the
entire address space is being replaced. With a bit of bad luck that
means the same address is still readable, and with even more bad luck
we'll read back a non-zero boolean. Thus, thinking the agent told us
it wanted to stay resident, we skip emitting the `uninjected` signal,
and DarwinHostSession has no idea that this happened. And then, on the
next request to attach to the same PID, we are stuck waiting forever
for the signal that never comes.

To solve this we allocate the data portion of the remote payload
locally, and create a shared mapping to where the data portion of the
remote payload is. This way we own that data, and are able to safely
examine it no matter what happens to the target process.

Thanks to @mrmacete for his awesome assistance in debugging and
brainstorming this issue.