Rework frida-pipe to simplify build systems
The underlying issue that we used to work around is that the Vala
compiler doesn't propagate the return value in case an error occurs,
which means the return value is 0 instead of -1, and this is treated
as EOF instead of an error.

While this is something worth fixing in the Vala compiler, our specific
use-case is actually better off by having each implementation implement
the entire {Input,Output}Stream classes. This way they can implement
different interfaces, e.g. when the underlying handle is pollable.