Improve helper process teardown performance and safety
We have already closed the connection so the helper knows it can stop,
and since we're on UNIX we can obviously unlink the files at any time.

This improves performance on teardown, but also avoids relying on GLib's
signal watching logic, which becomes problematic when more than one copy
of Frida is loaded in the same process. The issue is that each will call
sigaction() with the same signal number, meaning that only one of them
actually gets the callback. It may sound weird to have more than one copy
of Frida in the same process, but this may happen in case two Node.js
modules require different versions of Frida's Node.js bindings.