Make i/macOS process enumeration more robust
So we avoid undefined behavior in case a process dies after we've listed
the processes but before we've resolved all of their names.

The undefined behavior has been observed to result in entries where the
name is the same as the previous alive process. It could obviously also
result in junk.