Add support for specifying script options (#238)
The only supported option for now is for explicitly picking a JavaScript
runtime. This means that enable_jit() is now deprecated, but works
better than before as there is no longer any restriction on when it may
be called. All it does now is change the default for scripts created
without specifying which runtime to use.