Add support for opening backend-specific channels
Allows connecting to TCP services on a tethered iOS or Android device,
as well as lockdown services on a tethered iOS device.

Our Fruity and Droidy backends already need to do these things for their
own needs, so exposing this functionality to applications is very cheap
footprint-wise, and eliminates external dependencies in some cases.