Extend gadget ABI to support signaling a result
For now only on i/macOS, where the presence of a “frida_gadget_config”
entry in the apple[] string vector signals that the constructor function
is passed an extra argument that receives the result.

The result is an “int” that receives a non-negative value on success,
and a negative value on error. The “listen” interaction emits the TCP
listening port being used, and the other interactions emit zero.