Improve modulate.py to handle overlapping symbols
There might be more than one symbol associated with a given constructor
or destructor address. So far this has only been observed with ELFs, in
which case symbols may have a size specified. Because only the actual
constructor/destructor symbol has a size, we can simply pick that symbol
to resolve each collision.