Implement a new frida-pipe strategy for i/macOS
Turns out that our previous strategy of directly setting up a Mach port
in the target process becomes problematic with guarded Mach ports.

So instead we register a Mach service that's globally visible, and have
the target process contact it to fetch its end of the socketpair.
We also check with the sandbox up front, and issue an extension token
if needed.

However, if we're unable to register the Mach service with launchd, we
fall back to our previous strategy.