Handle Android devices where jailed is unsupported
If we fail to contact a frida-server (rooted mode), and we also fail to
start our helper (non-rooted mode), then just try to contact a remote
frida-server/frida-gadget like we used to.

Also cache that failure, as it's unlikely that we'll be able to start
our helper on subsequent tries.