Rewrite macOS spawn gating to use DTrace
Now that kernel extensions are being phased out, it's time to try
something simpler.

The idea is to use a destructive DTrace probe on the getpid() syscall,
as it happens to be used by dyld very early on. It's also a system call
that doesn't seem to be super-busy on a typical system, so we can afford
a few false stops. We would avoid such false stops if we were to put probes
on execve() and posix_spawn(), but the latter would require a lot more
complexity to deal with, so we avoid that for now.

This solution is somewhat crude as we're spawning /usr/sbin/dtrace
instead of using libdtrace, but OTOH this means we don't need to worry
too much about future OS releases potentially breaking the API.