Revert "Implement a new frida-pipe strategy for i/macOS"
It's flawed as we run into sandbox issues with many targets.

However, I'm not even sure that complex strategy was needed for iOS 15
after all, as the Mach ports crossing task boundaries are not guarded.
I don't have access to a jailbroken iOS 15 system to actually test this
though, so we'll just have to wait and see if somebody files an issue.

This reverts commit 7e4ec0c2ba80216801104b726caa52e38b88ebe6.