script-engine: Only access Stalker from JS thread
To fix deadlock when Gadget is blocking the dynamic linker with its
internal lock(s) held while waiting for resume(). In that case we are
processing the JS MainContext while blocking, with network I/O handled
by the Gadget thread. Because Stalker may interact with the dynamic
linker during its class_init(), we must ensure that it happens from
the JS thread and not the network I/O thread.