darwin: Fix early instrumentation on iOS 16
Unlike on macOS we cannot rely on dyld symbols in the shared cache copy,
since most of its symbols there are redacted. However, since the version
on the filesystem is identical, as ensured by dyld itself, we can simply
rebase any addresses we already have. This is also how we do it in our
implementation for jailed iOS.