linux: Revise /proc/self/stat num_threads faking strategy
It is risky to hook read() process-wide, as some callers may block
indefinitely, and we may get stuck on teardown waiting for Interceptor
to finish rolling back its changes.

So instead, only apply this instrumentation on Android, and simply
rebind libart.so's read() import. This means we don't have to worry
about long-lived calls, and the added overhead is limited to ART's code.