test-agent: Move thread suspend awareness test to slow mode
For now, until the deadlock it reproduces has been fixed. What happens
is that the suspended thread holds Gum's dlmalloc lock in the moment
when it gets suspended. So when the thread that suspended it carries on
and calls a hooked function, it ends up needing to allocate memory, and
deadlocks trying to get the dlmalloc lock.