compat: Generate a proper depfile with all inputs
Turns out compile_commands.json only lists a subset of the inputs, so we
would end up generating an incomplete depfile, causing unreliable
incremental builds.

So to solve this, we instead query Ninja for the inputs of each target.
This also gives us better granularity, as previously all of the outputs
would have the same inputs, and would all get invalidated if only one of
them changed.